modules/upgates/requirements.txt
-e file:///Users/cward/Repos/neven_cz/modules/abra
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
blinker==1.9.0
build==1.2.2.post1
cachecontrol==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
cohere==5.13.12
colorama==0.4.6
crashtest==0.4.1
decorator==5.1.1
deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
duckdb==1.2.0
dulwich==0.22.7
eval-type-backport==0.2.2
executing==2.2.0
fastavro==1.10.0
fastjsonschema==2.21.1
filelock==3.17.0
flask==3.1.0
frozenlist==1.5.0
fsspec==2025.2.0
google-auth==2.38.0
googleapis-common-protos==1.66.0
griffe==1.5.6
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
importlib-metadata==8.5.0
iniconfig==2.0.0
installer==0.7.0
ipdb==0.13.13
ipython==8.32.0
itsdangerous==2.2.0
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.1.0
jedi==0.19.2
jinja2==3.1.5
jiter==0.8.2
jsonpath-python==1.0.6
keyring==25.6.0
logfire==3.5.3
logfire-api==3.5.3
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.5.0
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.2
openai==1.61.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pip==25.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==2.0.1
poetry-core==2.0.1
prompt-toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-ai==0.0.22
pydantic-ai-slim==0.0.22
pydantic-core==2.27.2
pydantic-graph==0.0.22
pygments==2.19.1
pyproject-hooks==1.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyyaml==6.0.2
rapidfuzz==3.12.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
schedule==1.2.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2025.1.15.22
types-requests==2.32.0.20241016
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
-e file:///Users/cward/Repos/neven_cz/modules/upgates
urllib3==2.3.0
virtualenv==20.29.1
wcwidth==0.2.13
werkzeug==3.1.3
wrapt==1.17.2
xattr==1.1.4
xmltodict==0.14.2
yarl==1.18.3
zipp==3.21.0
